Move fabric databases refresh to data access and remove special case in Explorer
